requirements.txt
fire
pillow
pilkit